Move Cardinality et al to IR module
- Selectivity
- Cost
- CostPerRow
- Multiplier
