+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/CardinalityCostModel.scala
@@ -22,6 +22,7 @@ package org.neo4j.cypher.internal.compiler.v3_2.planner.logical
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.Metrics._
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans._
 import org.neo4j.cypher.internal.frontend.v3_2.ast.{HasLabels, Property}
+import org.neo4j.cypher.internal.ir.v3_2.{Cardinality, Cost, CostPerRow, Multiplier}
 
 object CardinalityCostModel extends CostModel {
