+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/spi/GraphStatistics.scala
@@ -19,8 +19,8 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_2.spi
 
-import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.{Cardinality, Selectivity}
 import org.neo4j.cypher.internal.frontend.v3_2.{LabelId, PropertyKeyId, RelTypeId}
+import org.neo4j.cypher.internal.ir.v3_2.{Cardinality, Selectivity}
 
 object GraphStatistics {
   val DEFAULT_RANGE_SELECTIVITY          = Selectivity.of(0.3).get
