+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/StubbedLogicalPlanningConfiguration.scala
@@ -22,9 +22,9 @@ package org.neo4j.cypher.internal.compiler.v3_2.planner
 import org.neo4j.cypher.internal.frontend.v3_2.ast.{Expression, HasLabels}
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.Metrics.{CardinalityModel, QueryGraphCardinalityModel, QueryGraphSolverInput}
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans.{IdName, LogicalPlan}
-import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.{Cardinality, Cost, Selectivity}
 import org.neo4j.cypher.internal.compiler.v3_2.spi.GraphStatistics
-import org.neo4j.cypher.internal.frontend.v3_2.{SemanticTable, LabelId}
+import org.neo4j.cypher.internal.frontend.v3_2.{LabelId, SemanticTable}
+import org.neo4j.cypher.internal.ir.v3_2.{Cardinality, Cost, Selectivity}
 
 class StubbedLogicalPlanningConfiguration(parent: LogicalPlanningConfiguration)
   extends LogicalPlanningConfiguration with LogicalPlanningConfigurationAdHocSemanticTable {
