+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/ExpandPlanningIntegrationTest.scala
@@ -26,6 +26,7 @@ import org.neo4j.cypher.internal.compiler.v3_2.planner.{LogicalPlanningTestSuppo
 import org.neo4j.cypher.internal.frontend.v3_2.ast._
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.frontend.v3_2.{LabelId, PropertyKeyId, SemanticDirection}
+import org.neo4j.cypher.internal.ir.v3_2.Cardinality
 
 class ExpandPlanningIntegrationTest extends CypherFunSuite with LogicalPlanningTestSupport2 {
