+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/cardinality/SelectivityCombinerTest.scala
@@ -19,8 +19,8 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_2.planner.logical.cardinality
 
-import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.Selectivity
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
+import org.neo4j.cypher.internal.ir.v3_2.Selectivity
 
 class SelectivityCombinerTest extends CypherFunSuite {
