+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/plans/IdSeekLeafPlannerTest.scala
@@ -21,12 +21,12 @@ package org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans
 
 import org.mockito.Mockito._
 import org.neo4j.cypher.internal.compiler.v3_2.planner._
-import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.Cost
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.Metrics.QueryGraphSolverInput
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.steps.idSeekLeafPlanner
 import org.neo4j.cypher.internal.frontend.v3_2.ast._
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.frontend.v3_2.{RelTypeId, SemanticDirection}
+import org.neo4j.cypher.internal.ir.v3_2.Cost
 
 import scala.collection.mutable
