+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/steps/SelectPatternPredicatesTest.scala
@@ -28,6 +28,7 @@ import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans._
 import org.neo4j.cypher.internal.frontend.v3_2.ast._
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.frontend.v3_2.{SemanticDirection, SemanticTable}
+import org.neo4j.cypher.internal.ir.v3_2.Cardinality
 
 class SelectPatternPredicatesTest extends CypherFunSuite with LogicalPlanningTestSupport {
