+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/steps/SortSkipAndLimitTest.scala
@@ -21,10 +21,11 @@ package org.neo4j.cypher.internal.compiler.v3_2.planner.logical.steps
 
 import org.neo4j.cypher.internal.compiler.v3_2.planner._
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans._
-import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.{Ascending, SortDescription, Cardinality, LogicalPlanningContext}
+import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.{Ascending, LogicalPlanningContext, SortDescription}
 import org.neo4j.cypher.internal.frontend.v3_2.ast
 import org.neo4j.cypher.internal.frontend.v3_2.ast.{AscSortItem, PatternExpression}
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
+import org.neo4j.cypher.internal.ir.v3_2.Cardinality
 
 class SortSkipAndLimitTest extends CypherFunSuite with LogicalPlanningTestSupport {
